+++ test/routes/form.test.js
@@ -73,7 +73,7 @@ const userService = {
   getUserByUserId: jest.fn(),
 }
 
-const riskService = {
+const alertService = {
   getEscapeProfile: jest.fn(),
 }
 
@@ -82,7 +82,7 @@ const formRoute = createRouter({
   offendersService,
   userService,
   riskProfilerService,
-  riskService,
+  alertService,
   authenticationMiddleware,
 })
 
@@ -129,7 +129,7 @@ beforeEach(() => {
   riskProfilerService.getSecurityProfile.mockResolvedValue({})
   riskProfilerService.getViolenceProfile.mockResolvedValue({})
   riskProfilerService.getExtremismProfile.mockResolvedValue({})
-  riskService.getEscapeProfile.mockResolvedValue({})
+  alertService.getEscapeProfile.mockResolvedValue({})
   db.pool.connect = jest.fn()
   db.pool.connect.mockResolvedValue(mockTransactionalClient)
 })
@@ -1343,7 +1343,7 @@ describe('GET /ratings/extremism', () => {
 describe('GET /categoriser/review', () => {
   test('Should perform a merge with existing data when loading review', () => {
     // data that should be persisted
-    riskService.getEscapeProfile.mockResolvedValue({
+    alertService.getEscapeProfile.mockResolvedValue({
       flagA: 'B2345XY',
     })
     riskProfilerService.getViolenceProfile.mockResolvedValue({
@@ -2018,7 +2018,7 @@ describe('GET /ratings/escapeRating', () => {
       escapeListAlerts: [],
       riskType: 'ESCAPE',
     }
-    riskService.getEscapeProfile.mockResolvedValue(escapeProfile)
+    alertService.getEscapeProfile.mockResolvedValue(escapeProfile)
     return request(app)
       .get(`/ratings/escapeRating/12345`)
       .expect(200)
@@ -2067,7 +2067,7 @@ describe('GET /ratings/escapeRating', () => {
       escapeListAlerts: [],
       riskType: 'ESCAPE',
     }
-    riskService.getEscapeProfile.mockResolvedValue(escapeProfile)
+    alertService.getEscapeProfile.mockResolvedValue(escapeProfile)
 
     return request(app)
       .get(`/ratings/escapeRating/12345`)
